+++ packages/components/src/legacy-select/SelectListBox.tsx
@@ -33,7 +33,6 @@ const Option = ({ item }: OptionProps) => (
     {...item.value}
     textValue={item.textValue}
     aria-label={item.textValue}
-    hideSelectionIndicator
   >
     {({ isDisabled, isSelected, selectionMode }) => (
       <>