+++ packages/components/src/list-box/ListBoxSection.tsx
@@ -3,16 +3,19 @@ import {
   ListBoxSection as AriaListBoxSection,
   type ListBoxSectionProps as AriaListBoxSectionProps,
 } from 'react-aria-components'
-import type { ListBoxSectionElement } from './types'
 import { ListBoxHeader } from './ListBoxHeader'
 
-export interface ListBoxSectionProps<T extends ListBoxSectionElement>
+export interface ListBoxSectionProps<T extends object>
   extends Omit<AriaListBoxSectionProps<T>, 'children'> {
+  /**
+   * @deprecated since v15.2.0 please remove as soon as we drop ComboBoxSection.
+   * Also consider using only AriaListBoxSectionProps, the custom children is just for the use of ListBoxHeader.
+   */
   name?: React.ReactNode
   children?: React.ReactNode
 }
 
-export const ListBoxSection = <T extends ListBoxSectionElement>({
+export const ListBoxSection = <T extends object>({
   children,
   name,
   ...rest